Fixing Dockerfile and DateTime format
