+++ Dockerfile
@@ -1,53 +1,50 @@
-# syntax = docker/dockerfile:1
-
 # Make sure RUBY_VERSION matches the Ruby version in .ruby-version
 ARG RUBY_VERSION=3.2.3
-FROM registry.docker.com/library/ruby:$RUBY_VERSION-slim as base
 
-# Rails app lives here
-WORKDIR /rails
+FROM docker.io/ruby:${RUBY_VERSION}-bookworm
 
-# Install base packages
-RUN apt-get update -qq && \
-    apt-get install --no-install-recommends -y \
-    curl \
-    libjemalloc2 \
-    libvips \
-    postgresql-client
-
-# Set production environment
-ENV RAILS_ENV="production" \
-    BUNDLE_DEPLOYMENT="1" \
-    BUNDLE_PATH="/usr/local/bundle" \
-    BUNDLE_WITHOUT="development"
-
-# Throw-away build stage to reduce size of final image
-FROM base as build
-
-# Install packages needed to build gems and node modules
-RUN apt-get install --no-install-recommends -y \
-  build-essential \
-    git \
-    libpq-dev \
-    node-gyp \
-    pkg-config \
-    python-is-python3
+ENV LANG C.UTF-8
 
 RUN set -eus; \
+    apt-get update -qq; \
     apt-get install -y --no-install-recommends \
-    shared-mime-info \
-    libpq-dev \
-    net-tools \
-    netcat \
+    build-essential \
+    git-all \
     htop \
-    strace
+    iputils-ping \
+    libjemalloc2 \
+    libpq-dev \
+    libvips \
+    libxml2-dev \
+    libxslt1-dev \
+    net-tools \
+    node-gyp \
+    nodejs \
+    pg-activity \
+    pkg-config \
+    postgresql-client \
+    python-is-python3 \
+    shared-mime-info \
+    strace \
+    ; \
+    apt-get clean; \
+    rm -rf /var/cache/apt/archives/* /var/lib/apt/lists/*;
 
-# Install JavaScript dependencies
-ARG NODE_VERSION=20.12.1
-ARG YARN_VERSION=4.1.1
-ENV PATH=/usr/local/node/bin:$PATH
+WORKDIR /rails
+
+# @see https://engineering.binti.com/jemalloc-with-ruby-and-docker/
+
+# ENV LD_PRELOAD=/usr/lib/x86_64-linux-gnu/libjemalloc.so.2 \
+ENV NODE_VERSION=20.12.1 \
+    YARN_VERSION=latest \
+    PATH=/usr/local/node/bin:$PATH
+
+RUN gem update --system -N
+
 RUN curl -sL https://github.com/nodenv/node-build/archive/master.tar.gz | tar xz -C /tmp/ && \
-    /tmp/node-build-master/bin/node-build "${NODE_VERSION}" /usr/local/node && \
+    /tmp/node-build-master/bin/node-build "${NODE_VERSION}" /usr/local/node
+
+RUN npm install -g npm@10.5.2 && \
     npm install -g yarn@$YARN_VERSION && \
     rm -rf /tmp/node-build-master
 
@@ -59,7 +56,7 @@ RUN bundle install && \
 
 # Install node modules
 COPY package.json yarn.lock ./
-RUN yarn install --frozen-lockfile
+RUN yarn install
 
 # Copy application code
 COPY . .
@@ -70,26 +67,25 @@ RUN bundle exec bootsnap precompile app/ lib/
 # Precompiling assets for production without requiring secret RAILS_MASTER_KEY
 RUN SECRET_KEY_BASE_DUMMY=1 ./bin/rails assets:precompile
 
-
-# Final stage for app image
-FROM base
-
 # Clean up installation packages to reduce image size
 RUN rm -rf /var/lib/apt/lists /var/cache/apt/archives
 
 # Copy built artifacts: gems, application
-COPY --from=build "${BUNDLE_PATH}" "${BUNDLE_PATH}"
-COPY --from=build /rails /rails
+COPY "${BUNDLE_PATH}" "${BUNDLE_PATH}"
 
 # Run and own only the runtime files as a non-root user for security
 RUN groupadd --system --gid 1000 rails && \
-    useradd rails --uid 1000 --gid 1000 --create-home --shell /bin/bash && \
-    chown -R rails:rails db log storage tmp
+    useradd rails --uid 1000 --gid 1000 --create-home --shell /bin/bash
+
+RUN chown -R rails:rails db log storage tmp /rails
+
 USER 1000:1000
+
+EXPOSE 3000
+
+RUN chmod 755 /rails/bin/docker-entrypoint
+
+ENV RAILS_ENV=production
 
 # Entrypoint prepares the database.
 ENTRYPOINT ["/rails/bin/docker-entrypoint"]
-
-# Start the server by default, this can be overwritten at runtime
-EXPOSE 3000
-CMD ["./bin/rails", "server"]
