+++ Gemfile
@@ -4,6 +4,7 @@ source 'https://rubygems.org'
 
 # Use main development branch of Rails
 gem 'rails', '=7.1.3.2'
+gem 'activesupport', '=7.1.3.2'
 
 # The original asset pipeline for Rails [https://github.com/rails/sprockets-rails]
 gem 'sprockets-rails'
