+++ app/controllers/time_slots_controller.rb
@@ -27,8 +27,8 @@ class TimeSlotsController < ApplicationController
   end
 
   def create
-    params[:time_slot][:start_time] = Time.from_picker(params.delete(:start_time))
-    params[:time_slot][:end_time] = Time.from_picker(params.delete(:end_time))
+    params[:time_slot][:start_time] = TimeHelper.from_picker(params.delete(:start_time))
+    params[:time_slot][:end_time] = TimeHelper.from_picker(params.delete(:end_time))
 
     @time_slot = TimeSlot.new(params[:time_slot])
 
@@ -40,8 +40,8 @@ class TimeSlotsController < ApplicationController
   end
 
   def update
-    params[:time_slot][:start_time] = Time.from_picker(params.delete(:start_time))
-    params[:time_slot][:end_time] = Time.from_picker(params.delete(:end_time))
+    params[:time_slot][:start_time] = TimeHelper.from_picker(params.delete(:start_time))
+    params[:time_slot][:end_time] = TimeHelper.from_picker(params.delete(:end_time))
 
     if @time_slot.update(params[:time_slot])
       redirect_to event_job_path(@event, @job)
