+++ bin/shchk
@@ -1,17 +1,21 @@
 #!/usr/bin/env bash
 # vim: ft=bash
 
-[[ -d ${HOME}/.bashmatic ]] || bash -c "$(curl -fsSL https://bashmatic.re1.re); bashmatic-install -q" ;
-source "${HOME}/.bashmatic/init.sh" 2>/dev/null
-
-command -v shellcheck >/dev/null || brew install shellcheck
+command -v shellcheck >/dev/null || {
+  if [[ $(uname -s) == "Darwin" ]]; then
+    brew install shellcheck
+  else
+    apt-get install shellcheck
+  fi
+}
 
 shellcheck -a --color -x bin/boot-up bin/start
 
 code=$?
 
 if [[ ${code} -eq 0 ]]; then
-  success "ShellCheck PASSED"
+  echo "ShellCheck PASSED ✅"
 else 
-  abort "ShellChecl FAILED"
+  echo "ShellChecl FAILED ❌"
+  exit 1
 fi
